Also move the SearchNodeComparator
