+++ src/Engine/PathCalcParams.java
@@ -1,10 +1,10 @@
 package Engine;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.Queue;
 
 import Engine.Utils.SearchNode;
-import Engine.Utils.SearchNodeComparator;
 import Terrain.GameMap;
 import Units.Unit;
 import Units.UnitContext;
@@ -173,6 +173,45 @@ public class PathCalcParams
         searchQueue.removeIf(node -> next.equals(node));
         searchQueue.add(new SearchNode(next, currentNode));
       }
+    }
+  }
+
+  /**
+   * Compares SearchNodes based on the amount of movePower they possess, and optionally
+   *   the remaining distance to a destination.
+   */
+  private static class SearchNodeComparator implements Comparator<SearchNode>
+  {
+    int[][] powerGrid;
+    private final boolean hasDestination;
+    private int xDest;
+    private int yDest;
+
+    public SearchNodeComparator(int[][] powerGrid)
+    {
+      this.powerGrid = powerGrid;
+      hasDestination = false;
+      xDest = 0;
+      yDest = 0;
+    }
+
+    public SearchNodeComparator(int[][] powerGrid, int x, int y)
+    {
+      this.powerGrid = powerGrid;
+      hasDestination = true;
+      xDest = x;
+      yDest = y;
+    }
+
+    @Override
+    public int compare(SearchNode o1, SearchNode o2)
+    {
+      int firstDist = Math.abs(o1.x - xDest) + Math.abs(o1.y - yDest);
+      int secondDist = Math.abs(o2.x - xDest) + Math.abs(o2.y - yDest);
+
+      int firstPowerEstimate = powerGrid[o1.x][o1.y] - ((hasDestination) ? firstDist : 0);
+      int secondPowerEstimate = powerGrid[o2.x][o2.y] - ((hasDestination) ? secondDist : 0);
+      return secondPowerEstimate - firstPowerEstimate;
     }
   }
 }
